Move all the other stuff out of views.py
