+++ core/views/__init__.py
@@ -15,3 +15,7 @@ from .export import *
 from .external_background_tasks import *
 from .analysis import *
 from .index import *
+from .global_messages import *
+from .search import *
+from .oai import *
+from .configuration import *
