+++ core/views/views.py
@@ -2,24 +2,14 @@
 from __future__ import unicode_literals
 
 # generic
-import json
 import logging
 import re
 
 # django
-from django.contrib.auth.decorators import login_required
 from django.core.urlresolvers import reverse
-from django.http import HttpResponse, JsonResponse
-from django.shortcuts import render, redirect
 
 # import models
 from core import models
-
-# import oai server
-from core.oai import OAIProvider
-
-# import mongo dependencies
-from core.mongo import *
 
 # Get an instance of a logger
 logger = logging.getLogger(__name__)
@@ -193,165 +183,3 @@ def breadcrumb_parser(request):
     # return
     return crumbs
 
-
-
-####################################################################
-# Configuration 												   #
-####################################################################
-
-@login_required
-def configuration(request):
-    # get all transformations
-    transformations = models.Transformation.objects.filter(use_as_include=False)
-
-    # get all OAI endpoints
-    oai_endpoints = models.OAIEndpoint.objects.all()
-
-    # get all validation scenarios
-    validation_scenarios = models.ValidationScenario.objects.all()
-
-    # get record identifier transformation scenarios
-    rits = models.RecordIdentifierTransformationScenario.objects.all()
-
-    # get all bulk downloads
-    bulk_downloads = models.DPLABulkDataDownload.objects.all()
-
-    # get field mappers
-    field_mappers = models.FieldMapper.objects.all()
-
-    # return
-    return render(request, 'core/configuration.html', {
-        'transformations': transformations,
-        'oai_endpoints': oai_endpoints,
-        'validation_scenarios': validation_scenarios,
-        'rits': rits,
-        'field_mappers': field_mappers,
-        'bulk_downloads': bulk_downloads,
-        'breadcrumbs': breadcrumb_parser(request)
-    })
-
-
-@login_required
-def oai_endpoint_payload(request, oai_endpoint_id):
-    """
-    Return JSON of saved OAI endpoint information
-    """
-
-    # retrieve OAIEndpoint
-    oai_endpoint = models.OAIEndpoint.objects.get(pk=oai_endpoint_id)
-
-    # pop state
-    oai_endpoint.__dict__.pop('_state')
-
-    # return as json
-    return JsonResponse(oai_endpoint.__dict__)
-
-
-@login_required
-def dpla_bulk_data_download(request):
-    """
-    View to support the downloading of DPLA bulk data
-    """
-
-    if request.method == 'GET':
-
-        # if S3 credentials set
-        if settings.AWS_ACCESS_KEY_ID and \
-                settings.AWS_SECRET_ACCESS_KEY and \
-                settings.AWS_ACCESS_KEY_ID is not None and \
-                settings.AWS_SECRET_ACCESS_KEY is not None:
-
-            # get DPLABulkDataClient and keys from DPLA bulk download
-            dbdc = models.DPLABulkDataClient()
-            bulk_data_keys = dbdc.retrieve_keys()
-
-        else:
-            bulk_data_keys = False
-
-        # return
-        return render(request, 'core/dpla_bulk_data_download.html', {
-            'bulk_data_keys': bulk_data_keys,
-            'breadcrumbs': breadcrumb_parser(request)
-        })
-
-    if request.method == 'POST':
-        # OLD ######################################################################
-        logger.debug('initiating bulk data download')
-
-        # get DPLABulkDataClient
-        dbdc = models.DPLABulkDataClient()
-
-        # initiate download
-        dbdc.download_and_index_bulk_data(request.POST.get('object_key', None))
-
-        # return to configuration screen
-        return redirect('configuration')
-
-
-####################################################################
-# OAI Server 													   #
-####################################################################
-
-def oai(request, subset=None):
-    """
-    Parse GET parameters, send to OAIProvider instance from oai.py
-    Return XML results
-    """
-
-    # get OAIProvider instance
-    op = OAIProvider(request.GET, subset=subset)
-
-    # return XML
-    return HttpResponse(op.generate_response(), content_type='text/xml')
-
-
-####################################################################
-# Global Search													   #
-####################################################################
-
-def search(request):
-    """
-    Global search of Records
-    """
-
-    # if search term present, use
-    q = request.GET.get('q', None)
-    if q:
-        search_params = json.dumps({'q': q})
-        logger.debug(search_params)
-    else:
-        search_params = None
-
-    # generate hierarchy_dict
-    job_hierarchy = _stateio_prepare_job_hierarchy()
-
-    return render(request, 'core/search.html', {
-        'search_string': q,
-        'search_params': search_params,
-        'job_hierarchy_json': json.dumps(job_hierarchy),
-        'breadcrumbs': breadcrumb_parser(request),
-        'page_title': ' | Search'
-    })
-
-
-####################################################################
-# Global Messages												   #
-####################################################################
-
-@login_required
-def gm_delete(request):
-    if request.method == 'POST':
-        # get gm_id
-        gm_id = request.POST.get('gm_id')
-
-        # init GlobalMessageClient
-        gmc = models.GlobalMessageClient(request.session)
-
-        # delete by id
-        results = gmc.delete_gm(gm_id)
-
-        # redirect
-        return JsonResponse({
-            'gm_id': gm_id,
-            'num_removed': results
-        })
